Move TreeHasher to toplevel package, use underscore test package
