+++ defaulthasher/hasher_test.go
